requirements.txt
# Base
torch==1.10.0
torchvision==0.11.1
pytorch-lightning==1.5.1
hydra-core==1.1.1


# Tests
pytest==6.2.5